// orbito-frontend/src/core/api/generated/models/subscriptionDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Orbito API
 * OpenAPI spec version: v1
 */

export interface SubscriptionDto {
  id?: string;
  tenantId?: string;
  clientId?: string;
  planId?: string;
  /** @nullable */
  status?: string | null;
  amount?: number;
  /** @nullable */
  currency?: string | null;
  billingPeriodValue?: number;
  /** @nullable */
  billingPeriodType?: string | null;
  startDate?: string;
  /** @nullable */
  endDate?: string | null;
  nextBillingDate?: string;
  isInTrial?: boolean;
  /** @nullable */
  trialEndDate?: string | null;
  /** @nullable */
  externalSubscriptionId?: string | null;
  createdAt?: string;
  /** @nullable */
  cancelledAt?: string | null;
  /** @nullable */
  updatedAt?: string | null;
  /** @nullable */
  clientCompanyName?: string | null;
  /** @nullable */
  clientEmail?: string | null;
  /** @nullable */
  clientFirstName?: string | null;
  /** @nullable */
  clientLastName?: string | null;
  /** @nullable */
  planName?: string | null;
  /** @nullable */
  planDescription?: string | null;
  paymentCount?: number;
  totalPaid?: number;
  /** @nullable */
  lastPaymentDate?: string | null;
}